% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments_chen.R
\name{chen_kmoment}
\alias{chen_kmoment}
\title{The k-th raw moment of the Chen distribution}
\usage{
chen_kmoment(k, lambda, gamma)
}
\arguments{
\item{k}{a positive integer.}

\item{lambda, gamma}{parameter values > 0.}
}
\value{
Estimated value of k-th raw moment, based on numerical integration,
as well as integral absolute error obtained from the function integrate.
}
\description{
Computes the k-th raw moment of the Chen distribution.
}
\details{
To obtain the value of k-th raw moment of the Chen distribution, it is
necessary to use numerical integration. For that purpose, the R function
'integrate()' can be used, which returns the estimated value of the integral
and also the integral absolute error (see details about function
'integrate()'). Therefore, to obtain the variance, the first component of
each k-th raw moments must be selected.
}
\examples{
chen_kmoment(k = 1, lambda = .1, gamma = .5) # First raw moment.
chen_kmoment(k = 2, lambda = .1, gamma = .5) # Second raw moment.
chen_kmoment(k = 2, lambda = .1, gamma = .5)[1] -
chen_kmoment(k = 1, lambda = .1, gamma = .5)[1]^2 # Variance.

}
