% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrl_chen.R
\name{chen_mrl}
\alias{chen_mrl}
\title{Mean residual life function of the Chen distribution}
\usage{
chen_mrl(x, lambda, gamma)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda, gamma}{parameter values > 0.}
}
\value{
Estimated value of mean residual life function, based on numerical
integration.
}
\description{
Computes the mean residual life function of the Chen distribution.
}
\examples{
chen_mrl(x = 5, lambda = .1, gamma = .5)

}
