% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chen.R
\name{chen_plot}
\alias{chen_plot}
\title{Plot function of the Chen distribution}
\usage{
chen_plot(
  data_type,
  from = NULL,
  to = NULL,
  xlim = NULL,
  ylim = NULL,
  x = NULL,
  lambda,
  gamma,
  log = FALSE,
  func_type,
  title,
  col = "blue",
  lty = 1
)
}
\arguments{
\item{data_type}{specifies whether the input is a x vector of data values or
an expression. Possible types are

- 'data' for data values,

- 'expression' for expression.}

\item{from}{lower x axis limit, by default from = 0.}

\item{to}{upper x axis limit, by default to = 1.}

\item{xlim}{x axis limits, by default xlim = c(from, to).}

\item{ylim}{y axis limits.}

\item{x}{vector of data values when data_type = "data".}

\item{lambda, gamma}{parameter values > 0.}

\item{log}{logical value.}

\item{func_type}{specifies the type of function to be plotted. Possible
types are

- 'density' for density plot,

- 'hazard' for hazard plot,

- 'cumulative hazard' for cumulative hazard plot,

- 'survival' for survival plot,

- 'cumulative distribution' for cumulative distribution plot,

- 'quantile' for quantile plot.}

\item{title}{title of the graphic.}

\item{col}{to set the color of the graphic.}

\item{lty}{to set the line type.}
}
\value{
If log = TRUE, numeric value of the logarithm of the function.

If cum_haz = TRUE, numeric value of the cumulative hazard function.

graphic of the chosen chen function.
}
\description{
Plots the density, cumulative distribution, hazard, cumulative hazard,
survival and quantile functions of the Chen distribution.
}
\examples{

#  Example of plotting cumulative distribution using an expression

chen_plot(data_type = "expression", from = 0, to = 6, lambda = 2,
gamma = 0.3, func_type = "cumulative distribution",
title = "Chen cumulative distribution function (expression)")

#  Example of plotting a bathtub-shaped hazard function using an expression

chen_plot(data_type = "expression", lambda = 2, gamma = 0.3,
func_type = "hazard", title = "Chen hazard function (expression)", to = 100,
ylim = c(0.8, 1.8))

# Example of plotting a bathtub-shaped hazard function using data points

x_data <- seq(0.0000001, 100, by = 0.1)
chen_plot(data_type = "data", x = x_data, lambda = 2, gamma = 0.3,
func_type = "hazard", title = "Chen hazard function (data points)",
to = 100, ylim = c(0.8, 1.8))

}
