% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_chen.R
\name{chen_shape}
\alias{chen_shape}
\title{Bowley skewness and Moors kurtosis of the Chen distribution}
\usage{
chen_shape(lambda, gamma, measure)
}
\arguments{
\item{lambda, gamma}{parameter values > 0.}

\item{measure}{Specifies the type of measure to be computed. Possible
types are:

 - 'bowley' for Bowley skewness,
 - 'moors' for Moors kurtosis.}
}
\value{
Numeric value of the Bowley skewness or the Moors kurtosis.
}
\description{
Computes robust skewness and kurtosis measures based on quantile function of
the Chen distribution.
}
\details{
The Chen distribution has no closed-form expression for the moments.
Therefore, the classical measures of skewness and kurtosis based on moments
are intractable. In such cases, quantile-based measures are often considered,
namely the Bowley skewness and Moors kurtosis.
}
\examples{
chen_shape(2, 0.3, measure = 'bowley') # Bowley skewness

}
