% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dchen.R
\name{dchen}
\alias{dchen}
\title{Density function of the Chen distribution}
\usage{
dchen(x, lambda, gamma, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda, gamma}{parameter values > 0.}

\item{log}{logical value}
}
\value{
Numeric value of the density function.

If log = TRUE, numeric value of the logarithm of the function.
}
\description{
Compute the density function of the Chen distribution.
}
\examples{
dchen(2, 1, 1, log = FALSE) # density function

}
