% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decp.R, R/global.R
\name{decp}
\alias{decp}
\title{Density function}
\usage{
decp(x, lambda, gamma, phi, log = FALSE)

decp(x, lambda, gamma, phi, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda, gamma}{parameter values > 0.}

\item{phi}{parameter value > 0 or < 0.}

\item{log}{logical value.}
}
\value{
Numeric value of the density function.

If log = TRUE, numeric value of the logarithm of the function.

Numeric value of the density function.

If log = TRUE, numeric value of the logarithm of the function.
}
\description{
Compute the density function of the extended Chen-Poisson (ecp) distribution.

Compute the extended Chen-Poisson (ecp) density function.
}
\examples{
decp(2, 1, 1, 1, log = FALSE) # density function

decp(2, 1, 1, 1, log = FALSE) # density function

}
