% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecp.R
\name{ecp_plot}
\alias{ecp_plot}
\title{Plot function}
\usage{
ecp_plot(
  data_type,
  from = NULL,
  to = NULL,
  xlim = NULL,
  ylim = NULL,
  x = NULL,
  lambda = 6,
  gamma = 0.3,
  phi = 30,
  log = FALSE,
  func_type,
  title,
  col = "black",
  lty = 1
)
}
\arguments{
\item{data_type}{specifies whether the input is a x vector of data values or
an expression.}

\item{from}{lower x axis limit.}

\item{to}{upper x axis limit.}

\item{xlim}{x axis limits.}

\item{ylim}{y axis limits.}

\item{x}{vector of data values.}

\item{lambda, gamma}{parameter values > 0.}

\item{phi}{parameter value > 0 or < 0.}

\item{log}{logical value.}

\item{func_type}{specifies the type of function to be plotted.}

\item{title}{title of the graphic.}

\item{col}{to set the color of the graphic.}

\item{lty}{to set the line type.}
}
\value{
If log = TRUE, numeric value of the logarithm of the function.

If cum_haz = TRUE, numeric value of the cumulative hazard function.

graphic of the chosen ecp function.
}
\description{
Plots the density, cumulative distribution, hazard, cumulative hazard,
survival and quantile functions of the extended Chen-Poisson (ecp)
distribution.
}
\examples{

ecp_plot(data_type = "expression", from = 0, to = 1,
func_type = "cumulative distribution",
title = "Cumulative Distribution Function (Expression)")
#  Example of plotting cumulative distribution using an expression
}
