% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecp.R
\name{hecp}
\alias{hecp}
\title{Hazard function}
\usage{
hecp(x, lambda, gamma, phi, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda, gamma}{parameter values > 0.}

\item{phi}{parameter value != 0.}

\item{log}{logical value}
}
\value{
Numeric value of the hazard function.

If log = TRUE, numeric value of the logarithm of the function.
}
\description{
Compute the hazard function of the extended Chen-Poisson (ecp) distribution.
}
\examples{
hecp(2, 1, 1, 1, log = FALSE) # hazard function

}
