% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchen.R
\name{pchen}
\alias{pchen}
\title{Distribution function of the Chen distribution}
\usage{
pchen(q, lambda, gamma, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{lambda, gamma}{parameter values > 0.}

\item{lower_tail}{similar to lower.tail}

\item{log_p}{logical value}
}
\value{
Numeric value of the distribution function.

If lower_tail = FALSE, numeric value of the survival function.

If log_p = TRUE, numeric value of the logarithm of the function.
}
\description{
Compute the cumulative distribution function of the Chen distribution.
}
\examples{
pchen(2, 1, 1, lower_tail = TRUE, log_p = FALSE) # distribution function
pchen(2, 1, 1, lower_tail = FALSE, log_p = FALSE) # survival function

}
