% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/qecp.R
\name{qecp}
\alias{qecp}
\title{Quantile function}
\usage{
qecp(p, lambda, gamma, phi, lower_tail = TRUE, log_p = FALSE)

qecp(p, lambda, gamma, phi, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{lambda, gamma}{parameter values > 0.}

\item{phi}{parameter value > 0 or < 0.}

\item{lower_tail}{similar to lower.tail.}

\item{log_p}{logical value.}
}
\value{
Numeric value of the quantile function.

If lower_tail = FALSE, numeric value of
the logarithm of the function.

If log_p = TRUE, numeric value of the logarithm of the function.

Numeric value of the quantile function.

If lower_tail = FALSE, numeric value of
the logarithm of the function.

If log_p = TRUE, numeric value of the logarithm of the function.
}
\description{
Compute the extended Chen-Poisson (ecp) quantile function.

Compute the extended Chen-Poisson (ecp) quantile function.
}
\examples{
qecp(0.5, 2, 1, 1, lower_tail = TRUE, log_p = FALSE)
# quantile function

qecp(0.5, 2, 1, 1, lower_tail = TRUE, log_p = FALSE)
# quantile function

}
