% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchen.R
\name{rchen}
\alias{rchen}
\title{Function to generate a pseudo-random sample of the Chen distribution}
\usage{
rchen(n, lambda, gamma)
}
\arguments{
\item{n}{sample size.}

\item{lambda, gamma}{parameter values > 0.}
}
\value{
A vector of randomly generated numbers from the
extended Chen-Poisson distribution.
}
\description{
Generate a pseudo-random sample, without censoring, from the Chen
distribution.
}
\examples{
rchen(10,1,1) # random sample of size 10

}
