% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recp.R
\name{recp}
\alias{recp}
\title{Function to generate a pseudo-random sample}
\usage{
recp(n, lambda, gamma, phi)
}
\arguments{
\item{n}{sample size.}

\item{lambda, gamma}{parameter values > 0.}

\item{phi}{parameter value != 0.}
}
\value{
A vector of randomly generated numbers from the
extended Chen-Poisson distribution.
}
\description{
Generate a pseudo-random sample, without censoring, from the
extended Chen-Poisson (ecp) distribution.
}
\examples{
recp(10,1,1,1) # random sample of size 10

}
